lib/Redis/ScriptCache.pm: Preserve context-aware evalsha
Redis.pm implements evalsha in a context-aware manner using wantarray.
We want to preserve that behaviour, and the easiest way to do that is by
preserving this code the way it was originally written. Redis.pm
implements all the error-catching here anyway, the eval was superfluous.

Add a test-case for this too.